Split out merge_imports module from helpers::insert_use
